Moved attributes to a new module.
